refactor: move secret instance config methods to separate file
