+++ internal/secret/secret_instance_config.go
@@ -42,7 +42,7 @@ func (c *SecretInstanceConfig) getSecretInstanceOperations() *util.Operations {
 
 	// append secret instance operations
 	operations.AppendOperation(util.Operation{
-		Name: "secret object",
+		Name:   "secret object",
 		Create: c.createSecretObject,
 		Delete: c.deleteSecretObject,
 	})
@@ -175,4 +175,219 @@ func (c *SecretInstanceConfig) createSecretObject() error {
 func (c *SecretInstanceConfig) deleteSecretObject() error {
 
 	return nil
+}
+
+// getSecretInstanceWorkloadTypeAndID returns the workload type and ID
+// for a secret instance.
+func getSecretInstanceWorkloadTypeAndId(secretInstance *v0.SecretInstance) (string, *uint, error) {
+	var workloadInstanceType string
+	var workloadInstanceID *uint
+
+	switch {
+	case secretInstance.WorkloadInstanceID != nil:
+		workloadInstanceType = util.TypeName(v0.WorkloadInstance{})
+		workloadInstanceID = secretInstance.WorkloadInstanceID
+	case secretInstance.HelmWorkloadInstanceID != nil:
+		workloadInstanceType = util.TypeName(v0.HelmWorkloadInstance{})
+		workloadInstanceID = secretInstance.HelmWorkloadInstanceID
+	default:
+		return "", nil, errors.New("secret instance must be attached to a workload instance or a helm workload instance")
+	}
+
+	return workloadInstanceType, workloadInstanceID, nil
+}
+
+// getThreeportobjects returns all threeport objects required for
+// gateway instance reconciliation
+func (c *SecretInstanceConfig) getThreeportObjects() error {
+	var err error
+
+	// get kubernetes runtime instance
+	c.kubernetesRuntimeInstance, err = client.GetKubernetesRuntimeInstanceByID(
+		c.r.APIClient,
+		c.r.APIServer,
+		*c.secretInstance.KubernetesRuntimeInstanceID,
+	)
+	if err != nil {
+		return fmt.Errorf("failed to get kubernetes runtime instance: %w", err)
+	}
+
+	// get kubernetes runtime definition
+	c.kubernetesRuntimeDefinition, err = client.GetKubernetesRuntimeDefinitionByID(
+		c.r.APIClient,
+		c.r.APIServer,
+		*c.kubernetesRuntimeInstance.KubernetesRuntimeDefinitionID,
+	)
+	if err != nil {
+		return fmt.Errorf("failed to get kubernetes runtime definition: %w", err)
+	}
+
+	// get secret definition
+	c.secretDefinition, err = client.GetSecretDefinitionByID(
+		c.r.APIClient,
+		c.r.APIServer,
+		*c.secretInstance.SecretDefinitionID,
+	)
+	if err != nil {
+		return fmt.Errorf("failed to get gateway controller workload definition: %w", err)
+	}
+
+	// determine correct workload type and instance ID
+	c.workloadInstanceType, c.workloadInstanceId, err = getSecretInstanceWorkloadTypeAndId(c.secretInstance)
+	if err != nil {
+		return fmt.Errorf("failed to determine workload type and instance ID: %w", err)
+	}
+
+	// update secret config with correct workload instance
+	switch c.workloadInstanceType {
+	case util.TypeName(v0.WorkloadInstance{}):
+		// get workload instance
+		c.workloadInstance, err = client.GetWorkloadInstanceByID(
+			c.r.APIClient,
+			c.r.APIServer,
+			*c.secretInstance.WorkloadInstanceID,
+		)
+		if err != nil {
+			return fmt.Errorf("failed to get workload instance: %w", err)
+		}
+	case util.TypeName(v0.HelmWorkloadInstance{}):
+		// get helm workload instance
+		c.helmWorkloadInstance, err = client.GetHelmWorkloadInstanceByID(
+			c.r.APIClient,
+			c.r.APIServer,
+			*c.secretInstance.HelmWorkloadInstanceID,
+		)
+		if err != nil {
+			return fmt.Errorf("failed to get helm workload instance: %w", err)
+		}
+	default:
+		return errors.New("secret instance must be attached to a workload instance or a helm workload instance")
+	}
+
+	return nil
+}
+
+// validateThreeportState validates the state of the threeport objects
+// required for secret instance reconciliation
+func (c *SecretInstanceConfig) validateThreeportState() error {
+	// validate secret definition is reconciled
+	if !*c.secretDefinition.Reconciled {
+		return errors.New("secret definition not reconciled")
+	}
+
+	// validate workload is reconciled
+	switch c.workloadInstanceType {
+	case util.TypeName(v0.WorkloadInstance{}):
+		if !*c.workloadInstance.Reconciled {
+			return errors.New("workload instance not reconciled")
+		}
+	case util.TypeName(v0.HelmWorkloadInstance{}):
+		if !*c.helmWorkloadInstance.Reconciled {
+			return errors.New("helm workload instance not reconciled")
+		}
+	default:
+		return errors.New("secret instance must be attached to a workload instance or a helm workload instance")
+	}
+
+	// confirm secret controller is deployed
+	if err := c.confirmSecretControllerDeployed(); err != nil {
+		return fmt.Errorf("failed to confirm secret controller deployed: %w", err)
+	}
+
+	return nil
+}
+
+// confirmSecretControllerDeployed confirms the secret controller
+// is deployed
+func (c *SecretInstanceConfig) confirmSecretControllerDeployed() error {
+	if c.kubernetesRuntimeInstance.SecretsControllerInstanceID != nil {
+		workloadInstance, err := client.GetWorkloadInstanceByID(
+			c.r.APIClient,
+			c.r.APIServer,
+			*c.kubernetesRuntimeInstance.SecretsControllerInstanceID,
+		)
+		if err != nil {
+			return fmt.Errorf("failed to get secret controller workload instance: %w", err)
+		}
+		if !*workloadInstance.Reconciled {
+			return errors.New("secret controller not reconciled")
+		}
+
+		// secret controller is reconciled
+		return nil
+	}
+
+	resourceInventory, err := client.GetResourceInventoryByK8sRuntimeInst(
+		c.r.APIClient,
+		c.r.APIServer,
+		c.kubernetesRuntimeInstance.ID,
+	)
+	if err != nil {
+		return fmt.Errorf("failed to get dns management iam role arn: %w", err)
+	}
+
+	externalSecretsYaml, err := c.getExternalSecretsSupportServiceYaml(resourceInventory.SecretsManagerRole.RoleArn)
+	if err != nil {
+		return fmt.Errorf("failed to create external secrets: %w", err)
+	}
+
+	// create secret controller workload definition
+	workloadDefName := fmt.Sprintf("%s-%s", "external-secrets", *c.kubernetesRuntimeInstance.Name)
+	externalSecretsWorkloadDefinition := v0.WorkloadDefinition{
+		Definition:   v0.Definition{Name: &workloadDefName},
+		YAMLDocument: util.StringPtr(externalSecretsYaml),
+	}
+
+	// create secret controller workload definition
+	createdSecretControllerWorkloadDefinition, err := client.CreateWorkloadDefinition(
+		c.r.APIClient,
+		c.r.APIServer,
+		&externalSecretsWorkloadDefinition,
+	)
+	if err != nil {
+		return fmt.Errorf("failed to create secret controller workload definition: %w", err)
+	}
+
+	// create secret controller workload instance
+	createdSecretControllerWorkloadInstance, err := client.CreateWorkloadInstance(
+		c.r.APIClient,
+		c.r.APIServer,
+		&v0.WorkloadInstance{
+			Instance:                    v0.Instance{Name: &workloadDefName},
+			WorkloadDefinitionID:        createdSecretControllerWorkloadDefinition.ID,
+			KubernetesRuntimeInstanceID: c.kubernetesRuntimeInstance.ID,
+		},
+	)
+
+	// update kubernetes runtime instance with secret controller instance ID
+	c.kubernetesRuntimeInstance.SecretsControllerInstanceID = createdSecretControllerWorkloadInstance.ID
+	if _, err = client.UpdateKubernetesRuntimeInstance(
+		c.r.APIClient,
+		c.r.APIServer,
+		c.kubernetesRuntimeInstance,
+	); err != nil {
+		return fmt.Errorf("failed to update kubernetes runtime instance with secret controller instance ID: %w", err)
+	}
+
+	return nil
+}
+
+// configureSecretControllerJsonManifests configures the secret controller
+// json manifests
+func (c *SecretInstanceConfig) configureSecretControllerJsonManifests() ([]datatypes.JSON, error) {
+	var manifests []datatypes.JSON
+
+	secretStoreJson, err := c.getSecretStoreJson()
+	if err != nil {
+		return nil, fmt.Errorf("failed to get secret store: %w", err)
+	}
+	manifests = append(manifests, secretStoreJson)
+
+	externalSecretJson, err := c.getExternalSecretJson()
+	if err != nil {
+		return nil, fmt.Errorf("failed to marshal external secret: %w", err)
+	}
+	manifests = append(manifests, externalSecretJson)
+
+	return manifests, nil
 }
